+++ app/src/main/java/com/rizalanggoro/presensigo/presentation/pages/attendance/subject/index/SubjectAttendanceScreen.kt
@@ -1,4 +1,4 @@
-package com.rizalanggoro.presensigo.presentation.pages.attendance.subject
+package com.rizalanggoro.presensigo.presentation.pages.attendance.subject.index
 
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
