+++ app/src/main/java/com/rizalanggoro/presensigo/presentation/pages/attendance/subject/major/MajorScreen.kt
@@ -1,4 +1,4 @@
-package com.rizalanggoro.presensigo.presentation.pages.attendance.subject
+package com.rizalanggoro.presensigo.presentation.pages.attendance.subject.major
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
